confirm before delete configuration

+++ src/org/montsuqi/client/ConfigurationViewer.java
@@ -197,10 +197,13 @@ public class ConfigurationViewer{
 			public void actionPerformed(ActionEvent e) {
 				int selectedRow = clist.getSelectedRow();
 				if ( selectedRow >= 0 ) {
-					String configName = (String)(clist.getValueAt(selectedRow, 0));
-					DefaultTableModel model = (DefaultTableModel)clist.getModel();
-					conf.deleteConfiguration(configName);
-					updateConfigurationList(clist);
+					int result = JOptionPane.showConfirmDialog(f, Messages.getString("ConfigurationViewer.delete_confirm_message"), Messages.getString("ConfigurationViewer.delete_confirm"), JOptionPane.YES_NO_OPTION); //$NON-NLS-1$ //$NON-NLS-2$
+					if (result == JOptionPane.YES_OPTION) {
+						String configName = (String)(clist.getValueAt(selectedRow, 0));
+						DefaultTableModel model = (DefaultTableModel)clist.getModel();
+						conf.deleteConfiguration(configName);
+						updateConfigurationList(clist);
+					}
 				}
 			}
 		});
